+++ intsureview_fe/src/components/Form/form.tsx
@@ -1,4 +1,5 @@
 import React, { useState } from "react";
+import './Form.css';
 import {
   Button,
   FormControl,
@@ -11,7 +12,7 @@ import {
   Select,
   TextField,
 } from "@mui/material";
-import { postData, validateName, validatePhoneNumber } from "../api/formapi";
+import { postData, validateName, validatePhoneNumber } from "../../api/formapi";
 import FormErrorSnackbar from "./FormErrorSnackbar";
 import FormSuccessSnackbar from "./FormSuccessSnackbar";
